Scenes and Scripts for Changing Scenes
RayCast is used to detect the presence of the fridge/sandwich station. Temporary Godot guys are used as graphics indicating the position of the RayCast, Fridge, and Sandwich station. Use SPACEBAR to interact with fridge and sandwich station. A "Back to the Kitchen" button is used to go back to the kitchen from either scene (must be clicked with mouse)

Movement works with both WASD and arrow keys but is NOT the final movement, this is temporary movement made so I can get to the items I need to work with (included in Walker.gd). Please note only the arrow keys rotate the RayCast (however as this is temporary movement and not what will be implemented into the game I am not going to fix it).